+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Moved gorilla tracing to a separate package
+- Renamed `context` package to `fxcontext`
 
 
 ## [0.23.0] - 2018-04-29